Rename reader::data to reader::builtin_data
Prepares for the addition of an online_data module by giving the
existing builtin-dataset module a more specific name.
